Added several helper scripts for creating / destroying VMs and bootstrapping the K0s cluster itself.
The intent is for the create/destroy scripts to use environment
variables set by direnv or some similar approach. To make this
obvious, I'll write some short documentation on which variables
are available.

The install-k0s.sh script is just a quick-and-dirty way to create
a K0s cluster and get MetalLB + OpenEBS going for basic functionality.
It could probably use some better error checking, but it works well
enough for me, and the jury is out on whether or not anyone else is
interested.