Fixed issue where the repo existing would prevent install of package.
This affected Helm and MetalLB. Repos in my test env are always
installed, or at least only installed infrequently, so creating
a cluster always skipped installing these packages. I reorganized
the install process to split the checks, and also adjusted some
of the message output to be more consistent with the reorg.