Added Ansible sub-project.
I've had a "working" but unofficial Ansible setup for a few things,
but it could be more formalized. This is an effort to do so in an
accountable manner.

Initial commit only includes the working "common" and "firewall"
roles. Next will likely be "proxmox", and then once I convert my
systems to using nginx rather than lighttpd, the "webserver" role.