Added role for Proxmox management
This new role basically just sets up nginx properly to proxy the
web admin, and increases the frequency of sysstat collections for
better sar output. It may do more later, but this is basically the
bare minimum for a "new" proxmox server.

It may also make sense to install the helper scripts for managing
Kubernetes and other VMs and then remove them from the README.