Added a role to deploy HAProxy for a High Availability control plane while using K0s.
This should make it easier to set up K0s during the initial bootstrap
steps. This is especially handy since it really only needs to be done
once, assuming the same hosts are used for all of the build-up / teardown
of the Kubernetes stack.

Also modified the role documentation to make it a bit more readable.